+++ lib/src/SearchBar.dart
@@ -4,6 +4,7 @@ import 'dart:ui';
 import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
+import 'package:loader_search_bar/src/AutoActive.dart';
 import 'package:loader_search_bar/src/StateHolder.dart';
 
 import 'QuerySetLoader.dart';
@@ -35,16 +36,17 @@ class SearchBar extends StatefulWidget implements PreferredSizeWidget {
     this.loader,
     this.overlayStyle,
     this.searchHint = 'Tap to search...',
+    this.initialQuery,
+    this.controller,
     this.iconified = true,
     bool autofocus,
+    AutoActive autoActive,
     SearchItem searchItem,
-    ValueChanged<bool> onActivatedChanged,
     SearchBarAttrs attrs,
-  })  : this.autofocus = autofocus ?? iconified,
+  })  : this.autoActive = autoActive ?? AutoActive.off,
+        this.autofocus = autofocus ?? iconified,
         this.searchItem = searchItem ?? SearchItem.action(),
-        this.attrs = _initAttrs(iconified, attrs),
-        this.activatedChangedCallback =
-            onActivatedChanged ?? _blankActivatedCallback;
+        this.attrs = _initAttrs(iconified, attrs);
 
   static SearchBarAttrs _initAttrs(bool iconified, SearchBarAttrs attrs) {
     final defaultAttrs =
@@ -61,8 +63,6 @@ class SearchBar extends StatefulWidget implements PreferredSizeWidget {
     textBoxBackgroundColor: Colors.black12,
     textBoxOutlineColor: Colors.black26,
   );
-
-  static final ValueChanged<bool> _blankActivatedCallback = (_) {};
 
   /// Function being called whenever query changes with its current value
   /// as an argument.
@@ -87,6 +87,12 @@ class SearchBar extends StatefulWidget implements PreferredSizeWidget {
   /// Hint string being displayed until user inputs any text.
   final String searchHint;
 
+  /// Query value displayed for the first time in search field.
+  final String initialQuery;
+
+  /// Controller object allowing to access some properties of current state.
+  final SearchBarController controller;
+
   /// Indicating way of representing non-activated SearchBar:
   ///   true if widget should be showed as an action item in defaultAppBar,
   ///   false if widget should be merged with defaultAppBar.
@@ -95,9 +101,9 @@ class SearchBar extends StatefulWidget implements PreferredSizeWidget {
   /// Determining if search field should get focus once it becomes visible.
   final bool autofocus;
 
-  /// Callback function receiving widget's current state whenever user begins
-  /// or cancels/ends search action.
-  final ValueChanged<bool> activatedChangedCallback;
+  /// Allows to decide if SearchBar should be activated once it becomes
+  /// visible for the first time.
+  final AutoActive autoActive;
 
   /// Defining how to position and build search item widget in AppBar.
   final SearchItem searchItem;
@@ -129,39 +135,59 @@ class SearchBar extends StatefulWidget implements PreferredSizeWidget {
 class SearchBarState extends State<SearchBar> {
   static final _stateHolder = StateHolder<SearchBar, SearchBarState>();
 
-  final FocusNode searchFocusNode = FocusNode();
-
-  final TextEditingController queryInputController = TextEditingController();
-
-  QuerySetLoader get _safeLoader => widget.loader ?? QuerySetLoader.blank;
-
   bool activated = false;
-
-  bool focused = false;
-
-  bool queryNotEmpty = false;
 
   bool isClearingQuery = false;
 
   bool expanded;
 
+  bool _hasPendingScaffold = false;
+
+  FocusNode searchFocusNode = FocusNode();
+
+  TextEditingController queryInputController;
+
   String loaderQuery;
 
   Orientation currentOrientation;
 
+  bool get focused => searchFocusNode.hasFocus;
+
+  bool get queryNotEmpty => queryInputController.text.isNotEmpty;
+
+  QuerySetLoader get _safeLoader => widget.loader ?? QuerySetLoader.blank;
+
   EdgeInsets get screenPadding => MediaQuery.of(context).padding;
+
+  VoidCallback get clearQueryCallback =>
+      widget.controller?.onClearQuery ?? _onClearQuery;
+
+  VoidCallback get cancelSearchCallback =>
+      widget.controller?.onCancelSearch ?? _onCancelSearch;
 
   @override
   void initState() {
     super.initState();
     _stateHolder.add(this);
     expanded = !widget.iconified;
-    queryInputController.addListener(_onQueryControllerChange);
     searchFocusNode.addListener(_onSearchFocusChange);
+    _initSearchQuery();
+    _initAutoActive();
+  }
+
+  void _onSearchFocusChange() {
+    if (focused && !activated) {
+      setState(() => _updateActivated(true));
+    }
+  }
+
+  void _initSearchQuery() {
+    queryInputController?.dispose();
+    queryInputController = TextEditingController(text: widget.initialQuery)
+      ..addListener(_onQueryControllerChange);
   }
 
   void _onQueryControllerChange() {
-    queryNotEmpty = queryInputController.text.isNotEmpty;
     if (isClearingQuery) {
       isClearingQuery = false;
       onTextChange('');
@@ -169,81 +195,112 @@ class SearchBarState extends State<SearchBar> {
   }
 
   void onTextChange(String text) {
-    setState(() {
-      if (_safeLoader.loadOnEachChange) loaderQuery = text;
-    });
+    if (_safeLoader.loadOnEachChange) {
+      setState(() => loaderQuery = text);
+    }
     if (widget.onQueryChanged != null) widget.onQueryChanged(text);
   }
 
   void onTextSubmit(String text) {
-    setState(() {
-      if (!_safeLoader.loadOnEachChange) loaderQuery = text;
-    });
+    if (!_safeLoader.loadOnEachChange) {
+      setState(() => loaderQuery = text);
+    }
     if (widget.onQuerySubmitted != null) widget.onQuerySubmitted(text);
   }
 
-  void _onSearchFocusChange() {
+  void _initAutoActive() {
+    if (widget.autoActive.shouldActivate(this)) {
+      Future(() => _activate(false));
+    }
+  }
+
+  void onSearchAction() {
+    _initSearchQuery();
     setState(() {
-      focused = searchFocusNode.hasFocus;
-      if (focused && !activated) {
-        activated = true;
-        widget.activatedChangedCallback(true);
-      }
+      expanded = true;
+      _updateActivated(true);
     });
   }
 
-  void onCancelSearch() {
-    setState(() {
-      if (activated) {
-        activated = false;
-        widget.activatedChangedCallback(false);
-      }
-      if (widget.iconified) expanded = false;
-      loaderQuery = null;
-    });
-    queryInputController.clear();
+  void _onCancelSearch() {
     searchFocusNode.unfocus();
     widget.loader?.clearData();
-    _rebuildScaffold();
+    setState(() {
+      if (widget.iconified) expanded = false;
+      loaderQuery = null;
+      _updateActivated(false);
+    });
+  }
+
+  void _updateActivated(bool value) {
+    if (activated != value) {
+      activated = value;
+      widget.controller?.onActivatedChanged?.call(value);
+      if (activated) {
+        loaderQuery = queryInputController.text;
+      }
+      _rebuildScaffold();
+    }
+  }
+
+  void _onClearQuery() =>
+      queryNotEmpty ? _clearQueryField() : searchFocusNode.unfocus();
+
+  void _clearQueryField() {
+    isClearingQuery = true;
+    queryInputController.clear();
+    _requestSearchFocus();
+  }
+
+  void _activate(bool forceFocus) {
+    if (!activated) {
+      if (widget.iconified) {
+        onSearchAction();
+      } else {
+        _activateMerged();
+      }
+    }
+    if (activated && forceFocus && !focused) {
+      _focusSearchField();
+    }
+  }
+
+  void _activateMerged() {
+    if (widget.autofocus) {
+      _requestSearchFocus();
+    } else {
+      setState(() => _updateActivated(true));
+    }
+  }
+
+  void _focusSearchField() async {
+    await Future.doWhile(() => Future(() => _hasPendingScaffold));
+    _requestSearchFocus();
   }
 
   void _rebuildScaffold() {
-    Future.delayed(
-      Duration(milliseconds: 50),
-      () => Scaffold.of(context).setState(() {}),
-    );
+    if (!_hasPendingScaffold) {
+      _hasPendingScaffold = true;
+      Future.delayed(Duration(milliseconds: 50), () {
+        Scaffold.of(context).setState(() {});
+        _hasPendingScaffold = false;
+      });
+    }
   }
 
   void onPrefixSearchTap() {
-    FocusScope.of(context).requestFocus(searchFocusNode);
+    _requestSearchFocus();
     _highlightQueryText();
   }
+
+  void _requestSearchFocus() =>
+      FocusScope.of(context).requestFocus(searchFocusNode);
 
   void _highlightQueryText() {
     queryInputController.selection = TextSelection(
       baseOffset: queryInputController.value.text.length,
       extentOffset: 0,
     );
-  }
-
-  void onClearQuery() {
-    if (queryNotEmpty) {
-      _clearQueryField();
-    } else {
-      searchFocusNode.unfocus();
-    }
-  }
-
-  void _clearQueryField() {
-    isClearingQuery = true;
-    queryInputController.clear();
-  }
-
-  void onSearchAction() {
-    setState(() {
-      expanded = true;
-    });
-    _rebuildScaffold();
   }
 
   @override
@@ -257,7 +314,7 @@ class SearchBarState extends State<SearchBar> {
   Future<bool> onWillPop() {
     bool shouldPop;
     if (activated) {
-      onCancelSearch();
+      _onCancelSearch();
       shouldPop = false;
     } else {
       shouldPop = true;
@@ -266,5 +323,30 @@ class SearchBarState extends State<SearchBar> {
   }
 
   @override
-  Widget build(BuildContext context) => SearchBarBuilder(this, context);
+  Widget build(BuildContext context) {
+    widget.controller?._state = this;
+    return SearchBarBuilder(this, context);
+  }
+}
+
+class SearchBarController {
+  SearchBarController({
+    this.onCancelSearch,
+    this.onClearQuery,
+    this.onActivatedChanged,
+  });
+
+  final VoidCallback onCancelSearch;
+  final VoidCallback onClearQuery;
+  final ValueChanged<bool> onActivatedChanged;
+
+  SearchBarState _state;
+
+  void setQueryText(String text) => _state?.queryInputController?.text = text;
+  void startSearch({forceFocus = true}) => _state?._activate(forceFocus);
+  void cancelSearch() => _state?._onCancelSearch();
+  void clearQuery() => _state?._onClearQuery();
+  bool get isEmpty => _state != null ? !_state.queryNotEmpty : null;
+  bool get isActivated => _state.activated;
+  bool get isFocused => _state.focused;
 }